IB/mlx: Fix destroy flow of DC CNAK QP resources
Allocation of DC CNAK QP resources is done when driver is loaded but if
this operation fails it doesn't fail the initialization procedure of the
driver. So, when the driver is unloaded the cleanup of the DC CNAK QP
resources should consider the possibility that these were not allocated.

issue: 1192235
Change-Id: I2657991d07f8d43180d4ec5f8e00b92645d44dfb
Signed-off-by: Moni Shoua <monis@mellanox.com>