net/mlx5: Issue SW reset on FW assert
If a FW assert is considered fatal, indicated by a new bit in the health
buffer, reset the FW. After the reset go through the normal recovery
flow. Only one PF needs to issue the reset, so an attempt is made to prevent
the 2nd function from also issuing the reset. It's not an error if that
happens, it's just slows recovery.

issue: 1187431
Change-Id: Id5c72f130b8f06f02f85c93497e290e81ab921e2
Signed-off-by: Alex Vesker <valex@mellanox.com>
Signed-off-by: Daniel Jurgens <danielj@mellanox.com>