net/mlx5: Add resiliency for allocations on numas with no memory
It is possible for the HCA numa to be -1, which means the node has no
memory. In that case dma memory allocation should choose the nearest
memory numa available.

The problem is that some inbox kernels did not take the patch that fixes
this issue [1], so this patch provides a workaround in the driver.

[1] commit 82c1fc714763 ("mm: use numa_mem_id() in alloc_pages_node()")

Issue: 1275512
Change-Id: Icff8551b03ebdee14671f2fddfe62a2fc26d3fee
Signed-off-by: Gal Pressman <galp@mellanox.com>