Revert "net/mlx5: Best effort icm pages allocation"
This reverts commit 4ac2758df88f06a6de17f36955686e89d82562d1.

Current TO is 60sec with this mechanism we can do up to 2 page
allocation cycles after that we will just tiemout the init process.

The FW is not aware of the timeout and the fact that we released
the command interface now all sorts of issues (bad DMA addr) can
happen.

Reverting this because INIT_HCA fail is better than an NMI.
If we do want to use this mechanism it should be combined
with sw reset, if the command times out we will trigger a
sw reset which will release all the pages that were allocated.

Issue: 1314568
Issue: 1264360
Change-Id: I34f3e8f043df2ef4aa80709419f923c540440d0a
Signed-off-by: Alex Vesker <valex@mellanox.com>