net/mlx5: Add mutual exclusion mechanism for health sw reset
Since the FW can be shared between different PFs/VFs it is common
that more than one health poll will detected a failure, this can
lead to multiple resets which are unneeded and mutliple cr-dumps
which will highly increase the recovery time.

The solution is to use a FW locking mechanism using semaphore space
to provide a way to allow only one device to collect the cr-dump and
to issue a sw-reset.

issue: 1187431
Change-Id: I92b83af2b46f5be519ba34860149be14a216b7fe
Signed-off-by: Alex Vesker <valex@mellanox.com>