net/mlx4_core: Pass register/unregister requests per MAC to PF only once
Currently, each request by a VF to register a MAC address
(and to unregister the mac) is sent to the PF driver on the host.

According to the ConnectX-3 Programmer's Reference Manual (PRM),
the VF driver should track registration requests per mac per port,
and only send the register/unregister request to the PF once per MAC.
(See the ALLOC_RES command in the PRM).

This requirement is implemented by tracking duplicate
mac register/unregister requests using a reference counter.
Only the initial register request is passed on to the PF;
subsequent register requests for that mac return success
and increment a reference counter.

For unregistering a mac, all unregister requests on the VF result
in decrementing the reference counter. When that counter
becomes zero, the unregister request is passed on to the PF.

issue: 1310135
Change-Id: I38b042b24b171fc90f19fa5965add1356ab87c54
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>