net/mlx5e: Hold measurements for net_dim in a dedicated struct
Currently the general counters for the channel's statistics
are passed to net_dim as data samples. In order to have more control
over the data used for net_dim's algorithm, we now store the needed
data in a dedicated struct. This would be used in a downstream commit
to pass relevant data only to net_dim.

Issue:1296597
Change-Id: I282a935060af62a7370b0447ea479686861acadf
Signed-off-by: Tal Gilboa <talgi@mellanox.com>