IB/core: Introduce rdma_is_gid_attr_valid() to check GID attribute
Introduce an API rdma_is_gid_attr_valid() to check the validity of the
referenced GID attribute.
This is useful for users who have referenced to GID attribute but
wants to check at later point of time if the GID attribute has become
invalid or not before accessing GID attributes.

issue: 1293988
Change-Id: Ic162601fa5bbce864d1ae05f5fe17fd93a25ffad
Signed-off-by: Parav Pandit <parav@mellanox.com>