IB/core: Memory region registration latency improvement
Before this change, the memory registration flow determined if the
memory belongs to a peer device or to the local device by:
1- Call the peer clients callbacks and check if the memory belongs to
one of the peers
2- If not then the page belongs to the current process, therefore, call
get_user_pages()

Benchmarks with a high memory registration/un-registration rate may
suffer from a high latency because of this behaviour as it slows the
path of local memory registration.
To overcome this issue, the registration flow will treat the memory as a
local memory first, and only then, in case of failure, it will go to the
peers.

Issue: 1425818
Change-Id: Id748360bf833b8659355cd4af36d7b73d498581f
Signed-off-by: Feras Daoud <ferasda@mellanox.com>