net/mlx5e/mlx4_en: Fix min MTU validation
Kernel stack is passing SW MTU to the driver's ndo_change_mtu, treating
it as HW MTU and translating it via HW2SW_MTU is wrong, fix it by using
ETH_MIN_MTU definition for the comparison.

In addition, this patch makes the  necessary changes to the rdma.m4 and
compat in order to get ETH_MIN_MTU define for all kernels.
In future rebase, This patch should be merged with:
980e9c2d02ae ("net/mlx5e, net/mlx4_en: Add min/max checks for old
kernels")

Issue:1564785
Change-Id: Icb72ae4d0e816cd35d2ff485e171ad5dbc9cb0e6
Signed-off-by: Eran Ben Elisha <eranbe@mellanox.com>