net/mlx5: Get max number of VFs from PCI for BlueField
Supposedly, ECPF should get the max VFs supported by host PF through
querying FW. As this functionality is not ready in FW yet, we used a
hard coded number (127) as the max number of VFs.

However, there is a bug with flow steering for the second port if
driver creates peer missing rules for non existing VF resource.

As a workaround, still get the max VFs from PCI.

Issue: 1689283
Change-Id: Ic8187fe08047ede9de7f24924fa4973e78f94db9
Signed-off-by: Bodong Wang <bodong@mellanox.com>