net/mlx5: Fix duplicated offload rules issue
If user adds two same filters using skip_sw, the second command will
succeed. To fix it, if the matches are the same, we will compare the
filter handles. If the handles are the same, we consider it as
a replace command and continue. If the handles are not the same,
we consider it as a duplicate add and fail the command.

Issue: 1783221
Change-Id: I59ddfe4283ce45f4372078536982add515a704e4
Signed-off-by: Bodong Wang <bodong@mellanox.com>