RDMA/uverbs: Fix odp experimental sysfs interface
Upstream kernel (to upstream kernel 4.19-rc3)
commit c5c4d92e70f3 ("RDMA/uverbs: Use cdev_device_add() instead of
 cdev_add()") modified struct ib_uverbs_device to contain
struct device directly, instead of as a pointer.

We therefore need to replace calls to dev_get_drvdata() with
container_of() in all the odp uverbs syscalls.

This change was not done for mlnx_ofed 4.6 commit
(which is not upstream, and was taken as-is from mlnx_ofed 4.5):
commit 7694c1032904 ("IB/uverbs: Add EXP verbs prefetching ODP
MR pages")

issue: 1692774
Change-Id: Ibe5ebbb7141de80409cf9a019a8dfe6fd2efb767
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>