net/mlx5: Fix reading of counters if bulk reading isn't supported
If firmware returns capability log_max_flow_counter_bulk as 1,
it means the firmware doesn't support bulk. And to specify a query
of a single counter, we have to set the number of flow counters to read
as zero.

This also exposed an issue that fs counters bulk reading doesn't support
reading counters one by one if the counter id's aren't aligned,
which is fixed in this commit as well.

Issue: 1711841
Change-Id: I95c45eff747fdb28bfe254030fce818c19f4b1fc
Fixes: a351a1b03bf1 ('net/mlx5: Introduce bulk reading of flow counters')
Signed-off-by: Paul Blakey <paulb@mellanox.com>