net/mlx5: Eswitch, avoid using work queue on direct devlink invocation
If eswitch mode change is invoked directly using devlink (as opposed to
using the sysfs entry), call directly the mode set logic. This prevents
possible crashes caused by the handler accessing data provided as
arguments which may be freed by the caller.

The fix checks the caller name and acts according to that.

Issue: 1718496
Change-Id: I8d0c8df210d3f2852e3c873929a40a409c055857
Signed-off-by: Eli Cohen <eli@mellanox.com>