net/mlx5: Get max VF supported by host PF from firmware
For Bluefield, the driver on ARM side has no way to know the max VFs
supported from host PF. And it relies on a hard-coded value(63). The
problem is that host PF may have a different number configured. This
brings multiple problems for features relies on this number of max VF.
 1. Flow steering should not create flow rules for non-exist vport.
 2. Driver should not query any cap from non-exist vport.

After the support from firmware to query this value, driver no longer
need to define such a hard-coded value. And also eliminate the potential
failures.

Issue: 1748423
Change-Id: Ib349da4e4f73a2a00bbb795b7df6f7a0c18b3759
Signed-off-by: Bodong Wang <bodong@mellanox.com>