BACKPORTS: Fix location of the epoll event masks
The epoll event masks are defined in eventpoll.h in the upstream
kernel. But these events are defined in poll.h in the mofed
backports. This leads to redefinition errors at SCST compilation,
because SCST backports check for the events in eventpoll.h and
not in poll.h.

The epoll event masks have been moved to eventpoll.h, now this
corresponds to their location in the upstream kernel.

Issue: 1753150
Change-Id: I0c65c42925fb409bdfa2a4ebbb9173eacc206212
Signed-off-by: Sergey Gorenko <sergeygo@mellanox.com>