net/mlx5_core: RoCE LAG recovery flow optimization
In normal driver initialization flow with RoCE LAG:
1. Load the two ETH interfaces, and two IB interfaces.
2. LAG detects upper event and tears down the two IB interfaces.
3. LAG loads one IB interface, mlx5_bond_x that represents the bond netdevice.

In case we are already in RoCE LAG and going through recovery flow, we can avoid
loading the two IB interfaces and unloading them back, since we know that we are
in RoCE LAG mode. Instead, we load the mlx5_bond_0 IB interface directly.

This optimization saves around 3 seconds in the recovery flow.

issue: 1711914
Change-Id: I1863107301404387ac7a677a7e3cb426f100d523
Signed-off-by: Majd Dibbiny <majd@mellanox.com>