net/mlx5: Decrease vgroup vport counter only on VF removal
Since we only increment the vgroup vport counter only when adding
a VF to the group decrementing the counter should follow the same
logic and avoid decrementing the PF vport when removing it from
the group.

This fix will prevent a warning print the appears due to closing
a vgroup with a vport counter that is not zero.

Issue: 1759857
Fixes: 8eb99c417894 ("net/mlx5: Fix wrong counting of vgroup0 members")
Change-Id: I1645fc8307a95cf844a14c7d38b4a3b610c6efe7
Signed-off-by: Ariel Levkovich <lariel@mellanox.com>