vf-net-link-name.sh: Fix the issues of renaming representors
Rename REPs by udev according to uplink name slot or fallback to name
path if slot doesn't exists. The VF representor is still renamed as
{uplink name slot/path}_{vf_num}
This patch also fixes the issue caused by uplink REP.
Besides, for the case of uplink REP enslaved in bond, we need to get
phys_switch_id again, as the SWID of VF REP will be changed.

Issue: 1784907
Change-Id: I163d7d3eca615673bf32820da1e7c3e7ca314098
Signed-off-by: Jianbo Liu <jianbol@mellanox.com>
Reviewed-by: Roi Dayan <roid@mellanox.com>